+++ src/features/auth/LoginPage.tsx
@@ -1,8 +1,8 @@
-import { useAuth } from '@/features/auth/AuthContext';
 import { login } from '@/features/auth/auth';
 import { AxiosError } from 'axios';
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
+import { useAuth } from './auth';
 
 const LoginPage = () => {
   const [form, setForm] = useState({ email: '', password: '' });
@@ -44,7 +44,8 @@ const LoginPage = () => {
       console.error('지원하지 않는 소셜 로그인 제공자입니다.');
       return;
     }
-    const baseUrl = 'https://pickple.o-r.kr';
+    //const baseUrl = 'https://pickple.o-r.kr';
+    const baseUrl = 'http://localhost:8080'; // 개발 환경용
     window.location.href = `${baseUrl}/oauth2/authorization/${provider}`;
   };
